backend/requirements.txt
# Men's Circle Management Platform - Backend Requirements
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database and async ORM
sqlalchemy[asyncio]==2.0.23
psycopg2-binary==2.9.9
asyncpg==0.29.0
alembic==1.13.1

# Redis for caching and sessions
redis==5.0.1

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Payment processing
stripe==7.8.0

# External services
sendgrid==6.11.0
twilio==8.11.0

# Utilities
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.25.2

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1 